requirements.txt
absl-py==1.0.0
affine==2.4.0
altgraph==0.17.2
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
astunparse==1.6.3
attrs==21.4.0
Babel==2.9.1
babelfish==0.6.0
backcall==0.2.0
beautifulsoup4==4.4.1
bleach==4.1.0
brython==3.10.4
bs4==0.0.1
CacheControl==0.13.1
cachetools==5.0.0
camelot-py==0.10.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.9
click==8.0.4
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.4
computer==1.0.0
cryptography==36.0.2
cssselect==1.1.0
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
distlib==0.3.4
distro==1.7.0
Django==4.0.3
django-filter==23.2
django-php-bridge==0.1.1
djangorestframework==3.14.0
dnspython==2.2.1
easyocr==1.5.0
entrypoints==0.3
et-xmlfile==1.1.0
exceptiongroup==1.1.3
fake-useragent==0.1.11
fastapi==0.104.1
filelock==3.6.0
fire==0.4.0
firebase-admin==6.2.0
Flask==2.0.3
flatbuffers==2.0
fonttools==4.31.2
future==0.18.2
gast==0.5.3
gcloud==0.18.3
google-api-core==2.12.0
google-api-python-client==2.106.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.3
google-cloud-firestore==2.13.0
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.2
grpcio-status==1.59.2
gunicorn==20.1.0
h11==0.14.0
h5py==3.6.0
html-table-parser==0.1.0
html-table-parser-python3==0.2.0
httplib2==0.22.0
idna==3.3
imageio==2.19.5
importlib-metadata==4.11.3
imutils==0.5.4
iniconfig==1.1.1
insta-scrape==2.1.2
instagram-private-api==1.6.0.0
instascraper==2.0.4
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.1.0
jedi==0.18.1
Jinja2==3.0.3
json5==0.9.6
jsonschema==4.3.3
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyter-server==1.13.1
jupyterlab==3.2.5
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.2
jupyterlab-widgets==1.0.2
jws==0.1.3
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.0
libclang==14.0.1
lxml==4.8.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
msgpack==1.0.7
mysql-connector==2.2.9
mysql-connector-python==8.0.22
mysqlclient==2.1.0
nbclassic==0.3.4
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.8.5
notebook==6.4.6
numpy==1.22.2
oauth2client==4.1.3
oauthlib==3.2.0
opencv-python==4.6.0.66
opencv-python-headless==4.5.4.60
openpyxl==3.0.9
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parse==1.19.0
parso==0.8.3
pdfminer==20191125
pdfminer.six==20211012
pdfplumber==0.6.0
phpserialize==1.3
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.1
plotly==5.17.0
pluggy==1.0.0
prometheus-client==0.12.0
prompt-toolkit==3.0.24
proto-plus==1.22.3
protobuf==4.25.0
py==1.11.0
pyad==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.4.2
pydantic_core==2.10.1
PyDispatcher==2.0.5
pyee==8.2.2
Pygments==2.11.1
pyinstaller-hooks-contrib==2022.2
PyJWT==2.8.0
PyMySQL==1.0.2
pyparsing==3.0.6
PyPDF2==1.26.0
pypiwin32==223
pyppeteer==1.0.2
pyquery==1.4.3
Pyrebase==3.0.27
Pyrebase4==4.7.1
pyrsistent==0.18.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-jwt==2.0.1
python-multipart==0.0.6
pytz==2019.3
PyWavelets==1.3.0
pywin32==303
pywin32-ctypes==0.2.0
pywinpty==1.1.6
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.0
rasterio==1.3.8
requests==2.29.0
requests-html==0.10.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rsa==4.8
scikit-image==0.19.3
scipy==1.8.1
Send2Trash==1.8.0
setproctitle==1.2.3
six==1.13.0
sniffio==1.2.0
snuggs==1.4.7
soupsieve==2.3.1
sqlparse==0.3.0
starlette==0.27.0
tabula-py==2.3.0
tabulate==0.8.9
tenacity==8.2.3
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.25.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
tf-estimator-nightly==2.8.0.dev2021122109
tifffile==2022.5.4
tomli==2.0.1
torch==1.12.0
torchvision==0.13.0
tornado==6.1
tqdm==4.63.0
traitlets==5.1.1
twisted-iocpsupport==1.0.2
typing_extensions==4.8.0
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.7
uvicorn==0.24.0
validate-email==1.3
virtualenv==20.14.1
vision==1.0.0
vpn-slice==0.16.1
w3lib==1.22.0
Wand==0.6.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
websockets==10.2
Werkzeug==2.0.3
whitenoise==6.0.0
widgetsnbextension==3.5.2
word2number==1.1
wrapt==1.14.0
wxPython==4.2.0
zipp==3.7.0
